move alert data to separate file
