+++ pages/admin/alerts/@id/+data.ts
@@ -1,6 +1,5 @@
 import { PageContext } from "vike/types";
 import { JsonSerializable } from "@/shared/json-serializable";
-import { AlertData } from "@/server/data/alert/alert";
 import { App } from "@/server/app";
 import { nonNull, unique } from "@dan-schel/js-utils";
 import { DetailsError } from "@/server/alert-source/alert-source";
@@ -9,6 +8,7 @@ import { formatDate } from "@/server/data/disruption/period/utils/utils";
 import { ProcessingContextData } from "@/shared/types/processing-context-data";
 import { formatLineShapeNode } from "@/server/data/disruption/writeup/utils";
 import { AlertRepository } from "@/server/database-repository/alert-repository";
+import { AlertData } from "@/server/data/alert/alert-data";
 
 type UrlPreview = { html: string } | { error: string };
